+++ R/allGrid.R
@@ -1,37 +1,28 @@
 #' @rdname allUtils
 #' @order 3
-`allGrid` <- function(n, nperms, nr, nc, mirror, constant)
+`allGrid` <- function(n, nperms, nr, nc, mirror, constant,
+                      symmetric = FALSE)
 {
-    v <- seq_len(n)
-    X <- matrix(nrow = nperms, ncol = n)
-    idx <- 1
-    ## ncol == 2 is special case
-    if(nc == 2) {
-        X <- allSeries(n, nperms = nperms, mirror = mirror)
-    } else {
+    stopifnot(n == nr * nc)
+    orientations <- gridOrientations(nr, nc, mirror, symmetric)
+    expected <- n * nrow(orientations)
+    if(!isTRUE(all.equal(nperms, expected)))
+        stop("'nperms' does not match the number of grid permutations")
+    X <- matrix(nrow = expected, ncol = n)
+    idx <- 1L
+    for(k in seq_len(nrow(orientations))) {
         for(i in seq_len(nr)) {
             for(j in seq_len(nc)) {
-                ir <- seq(i, length = nr)%%nr
-                ic <- seq(j, length = nc)%%nc
-                ## block 1 - no reversals
+                ir <- seq(i, length = nr) %% nr
+                ic <- seq(j, length = nc) %% nc
+                if(orientations[k, "row"])
+                    ir <- rev(ir)
+                if(orientations[k, "col"])
+                    ic <- rev(ic)
                 X[idx, ] <- rep(ic, each = nr) * nr +
-                    rep(ir, len = nr * nc) + 1
-                if(mirror) {
-                    ## block 2 - rev rows but not columns
-                    X[idx + n, ] <- rep(ic, each = nr) * nr +
-                        rep(rev(ir), len = nr * nc) + 1
-                    ## block 3 - rev columns but not rows
-                    X[idx + (2*n), ] <- rep(rev(ic), each = nr) *
-                        nr + rep(ir, len = nr * nc) + 1
-                }
-                idx <- idx + 1
+                    rep(ir, length.out = nr * nc) + 1L
+                idx <- idx + 1L
             }
-        }
-        if(mirror) {
-            ## rev columns and rows
-            ## no calculations, just rev cols of block 1
-            v <- seq_len(n)
-            X[((3*n)+1):(4*n), ] <- X[v, rev(v)]
         }
     }
     X